Continued work on moving common code and specific switch implementation code
